MAINT: Move PyfftwBackend into the benchmark suite to avoid licensing issues
